# Clean up dbtest module