--- a/internal/dbtest/.gitignore
+++ b/internal/dbtest/.gitignore
@@ -1,0 +1,1 @@
+/bench*.txt