--- a/internal/dbtest/.snapshots/TestQuery-mysql-70
+++ b/internal/dbtest/.snapshots/TestQuery-mysql-70
@@ -1,1 +1,0 @@
-SELECT `soft_delete`.`id`, `soft_delete`.`deleted_at` FROM `soft_deletes` AS `soft_delete` WHERE `soft_delete`.`deleted_at` IS NOT NULL