--- a/internal/dbtest/db_test.go
+++ b/internal/dbtest/db_test.go
@@ -18,25 +18,53 @@ import (
 
 	_ "github.com/go-sql-driver/mysql"
 	_ "github.com/jackc/pgx/v4/stdlib"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 var ctx = context.TODO()
 
-func pg() *bun.DB {
+var allDBs = map[string]func(tb testing.TB) *bun.DB{
+	"pg":     pg,
+	"mysql":  mysql,
+	"sqlite": sqlite,
+}
+
+func pg(tb testing.TB) *bun.DB {
 	dsn := os.Getenv("PG")
 	if dsn == "" {
-		dsn = "postgres://postgres:@localhost:5432/test"
+		dsn = "postgres://postgres:postgres@localhost:5432/test?sslmode=disable"
 	}
 
 	sqldb := sql.OpenDB(pgdriver.NewConnector(pgdriver.WithDSN(dsn)))
-	db := bun.NewDB(sqldb, pgdialect.New())
-	return db
+	tb.Cleanup(func() {
+		assert.NoError(tb, sqldb.Close())
+	})
+
+	return bun.NewDB(sqldb, pgdialect.New())
 }
 
-func sqlite(t *testing.T) *bun.DB {
-	sqldb, err := sql.Open(sqliteshim.ShimName, "file::memory:?cache=shared")
-	require.NoError(t, err)
+func mysql(tb testing.TB) *bun.DB {
+	dsn := os.Getenv("MYSQL")
+	if dsn == "" {
+		dsn = "user:pass@/test"
+	}
+
+	sqldb, err := sql.Open("mysql", dsn)
+	require.NoError(tb, err)
+	tb.Cleanup(func() {
+		assert.NoError(tb, sqldb.Close())
+	})
+
+	return bun.NewDB(sqldb, mysqldialect.New())
+}
+
+func sqlite(tb testing.TB) *bun.DB {
+	sqldb, err := sql.Open(sqliteshim.DriverName(), "file::memory:?cache=shared")
+	require.NoError(tb, err)
+	tb.Cleanup(func() {
+		assert.NoError(tb, sqldb.Close())
+	})
 
 	sqldb.SetMaxIdleConns(1000)
 	sqldb.SetConnMaxLifetime(0)
@@ -44,32 +72,16 @@ func sqlite(t *testing.T) *bun.DB {
 	return bun.NewDB(sqldb, sqlitedialect.New())
 }
 
-func mysql(t *testing.T) *bun.DB {
-	dsn := os.Getenv("MYSQL")
-	if dsn == "" {
-		dsn = "root:pass@/test"
+func testEachDB(t *testing.T, f func(t *testing.T, db *bun.DB)) {
+	for name, newDB := range allDBs {
+		t.Run(name, func(t *testing.T) {
+			db := newDB(t)
+			if _, ok := os.LookupEnv("DEBUG"); ok {
+				db.AddQueryHook(bundebug.NewQueryHook(bundebug.WithVerbose()))
+			}
+			f(t, db)
+		})
 	}
-
-	sqldb, err := sql.Open("mysql", dsn)
-	require.NoError(t, err)
-
-	return bun.NewDB(sqldb, mysqldialect.New())
-}
-
-func dbs(t *testing.T) []*bun.DB {
-	dbs := []*bun.DB{
-		pg(),
-		sqlite(t),
-		mysql(t),
-	}
-
-	if false {
-		for _, db := range dbs {
-			db.AddQueryHook(bundebug.NewQueryHook(bundebug.WithVerbose()))
-		}
-	}
-
-	return dbs
 }
 
 func TestDB(t *testing.T) {
@@ -100,17 +112,13 @@ func TestDB(t *testing.T) {
 		{"testRunInTx", testRunInTx},
 	}
 
-	for _, db := range dbs(t) {
-		t.Run(db.Dialect().Name(), func(t *testing.T) {
-			defer db.Close()
-
-			for _, test := range tests {
-				t.Run(test.name, func(t *testing.T) {
-					test.run(t, db)
-				})
-			}
-		})
-	}
+	testEachDB(t, func(t *testing.T, db *bun.DB) {
+		for _, test := range tests {
+			t.Run(test.name, func(t *testing.T) {
+				test.run(t, db)
+			})
+		}
+	})
 }
 
 func testPing(t *testing.T, db *bun.DB) {