--- a/internal/dbtest/go.mod
+++ b/internal/dbtest/go.mod
@@ -33,4 +33,5 @@ require (
 	github.com/uptrace/bun/driver/pgdriver v0.1.0
 	github.com/uptrace/bun/driver/sqliteshim v0.0.0-00010101000000-000000000000
 	github.com/uptrace/bun/extra/bundebug v0.0.0-00010101000000-000000000000
+	modernc.org/sqlite v1.11.1
 )