--- a/internal/dbtest/soft_delete_test.go
+++ b/internal/dbtest/soft_delete_test.go
@@ -12,19 +12,7 @@ import (
 )
 
 func TestSoftDelete(t *testing.T) {
-	for _, db := range dbs(t) {
-		t.Run(db.Dialect().Name(), func(t *testing.T) {
-			defer db.Close()
-
-			_, err := db.NewDropTable().Model((*Video)(nil)).IfExists().Exec(ctx)
-			require.NoError(t, err)
-
-			_, err = db.NewCreateTable().Model((*Video)(nil)).Exec(ctx)
-			require.NoError(t, err)
-
-			testSoftDelete(t, db)
-		})
-	}
+	testEachDB(t, testSoftDelete)
 }
 
 type CustomTime struct {
@@ -55,10 +43,16 @@ type Video struct {
 }
 
 func testSoftDelete(t *testing.T, db *bun.DB) {
+	_, err := db.NewDropTable().Model((*Video)(nil)).IfExists().Exec(ctx)
+	require.NoError(t, err)
+
+	_, err = db.NewCreateTable().Model((*Video)(nil)).Exec(ctx)
+	require.NoError(t, err)
+
 	video1 := &Video{
 		ID: 1,
 	}
-	_, err := db.NewInsert().Model(video1).Exec(ctx)
+	_, err = db.NewInsert().Model(video1).Exec(ctx)
 	require.NoError(t, err)
 
 	// Count visible videos.