This PR cleans up dbtest module and also adds `cgobench.sh` script to compare Cgo vs pure Go performance. Note though that (afaik) Go test tooling doesn’t count unmanaged allocations in benchmarks.

<details>
<summary><code>./cgobench.sh</code></summary>

The results are from a single-core xhyve VM. They’d likely be different (and closer to reality) on a real multi-core hardware.

```
+ trap 'docker-compose down -v' EXIT
+ docker-compose down -v
Removing network dbtest_default
WARNING: Network dbtest_default not found.
+ docker-compose up -d
Creating network "dbtest_default" with the default driver
Creating dbtest_mysql_1    ... done
Creating dbtest_postgres_1 ... done
+ sleep 15
+ tee bench-purego.txt
+ CGO_ENABLED=0
+ go test '-run=^$' -bench=. -benchmem -count=10
goos: linux
goarch: amd64
pkg: github.com/uptrace/bun/internal/dbtest
cpu: Intel(R) Core(TM) i5-7500 CPU @ 3.40GHz
BenchmarkSelectOne/pg         	    3480	    401887 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3450	    337679 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3436	    402106 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3526	    341215 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3591	    344087 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3544	    340947 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3444	    342684 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3601	    336131 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3368	    344055 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3417	    395629 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/mysql      	    2920	    365868 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2905	    365219 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    3104	    376057 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    3081	    377231 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2938	    368638 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2980	    367282 ns/op	    5529 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2914	    370460 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2913	    444473 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    3045	    385622 ns/op	    5529 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2797	    502268 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/sqlite     	   20914	     79636 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   13849	     82900 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   28515	     47100 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   27804	     51868 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   25446	     45801 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   28273	     42094 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   28609	     41866 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   28209	     42520 ns/op	    5768 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   27694	     45552 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectOne/sqlite     	   27003	     45835 ns/op	    5752 B/op	      38 allocs/op
BenchmarkSelectSlice/mysql    	    1876	    595169 ns/op	   21600 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1918	    584965 ns/op	   21656 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    2011	    582370 ns/op	   21640 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1790	    590520 ns/op	   21656 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1910	    595561 ns/op	   21584 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1914	    581393 ns/op	   21616 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1898	    593710 ns/op	   21608 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1978	    569854 ns/op	   21608 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1965	    646123 ns/op	   21632 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1947	    594251 ns/op	   21584 B/op	     523 allocs/op
BenchmarkSelectSlice/sqlite   	    2185	    595697 ns/op	   57520 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2096	    592066 ns/op	   57568 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2234	    561429 ns/op	   57552 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2096	    551016 ns/op	   57680 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2143	    573733 ns/op	   57584 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2143	    543923 ns/op	   57488 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2148	    536199 ns/op	   57632 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2287	    536159 ns/op	   57600 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2112	    540763 ns/op	   57568 B/op	    1724 allocs/op
BenchmarkSelectSlice/sqlite   	    2149	    540052 ns/op	   57552 B/op	    1724 allocs/op
BenchmarkSelectSlice/pg       	    2030	    606481 ns/op	   18185 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    1923	    634502 ns/op	   18137 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2029	    683887 ns/op	   18185 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    1989	    610347 ns/op	   18169 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2035	    647988 ns/op	   18145 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    1888	    614395 ns/op	   18129 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    1963	    648966 ns/op	   18177 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    1999	    619779 ns/op	   18089 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    1892	    767474 ns/op	   18153 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    1894	    700270 ns/op	   18161 B/op	     415 allocs/op
BenchmarkSelectError/pg       	    3135	    373967 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3130	    437492 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3170	    374959 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3200	    406044 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3114	    373161 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3099	    396212 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3211	    452611 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    2745	    438284 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3122	    427090 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    2745	    419655 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/mysql    	    2773	    442368 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2910	    418477 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    3175	    438548 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2666	    423143 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    3112	    388114 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2913	    428283 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2895	    385445 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2929	    373715 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    3163	    430201 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2922	    375429 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/sqlite   	   79626	     14085 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   85053	     15196 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   84457	     15149 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   85706	     15189 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   82747	     14664 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   81600	     14388 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   72985	     14505 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   79443	     14628 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   85772	     14551 ns/op	    2640 B/op	      18 allocs/op
BenchmarkSelectError/sqlite   	   80599	     15629 ns/op	    2640 B/op	      18 allocs/op
PASS
ok  	github.com/uptrace/bun/internal/dbtest	574.141s
+ tee bench-cgo.txt
+ CGO_ENABLED=1
+ go test '-run=^$' -bench=. -benchmem -count=10 -tags cgosqlite
goos: linux
goarch: amd64
pkg: github.com/uptrace/bun/internal/dbtest
cpu: Intel(R) Core(TM) i5-7500 CPU @ 3.40GHz
BenchmarkSelectOne/pg         	    3235	    340826 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3489	    379595 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3580	    339394 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3576	    336511 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3470	    334943 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3588	    338340 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3564	    338760 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3603	    337985 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3561	    367017 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/pg         	    3607	    332421 ns/op	    5216 B/op	      16 allocs/op
BenchmarkSelectOne/mysql      	    3033	    361106 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    3075	    362322 ns/op	    5529 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2964	    399202 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2977	    369032 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    3093	    407869 ns/op	    5529 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    3240	    369133 ns/op	    5529 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2953	    395212 ns/op	    5529 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2979	    368225 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    2887	    366152 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/mysql      	    3090	    410558 ns/op	    5521 B/op	      25 allocs/op
BenchmarkSelectOne/sqlite     	   43356	     27579 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   41274	     29181 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   39752	     26789 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   42728	     28918 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   44907	     27153 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   43966	     29797 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   44887	     26899 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   44127	     27201 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   43742	     27097 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectOne/sqlite     	   45628	     28312 ns/op	    5704 B/op	      51 allocs/op
BenchmarkSelectSlice/pg       	    2040	    648919 ns/op	   18153 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2130	    598859 ns/op	   18145 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2126	    594248 ns/op	   18161 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2097	    600356 ns/op	   18137 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2115	    625712 ns/op	   18153 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2121	    594320 ns/op	   18129 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2115	    584872 ns/op	   18169 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2059	    580915 ns/op	   18129 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2104	    595978 ns/op	   18153 B/op	     415 allocs/op
BenchmarkSelectSlice/pg       	    2061	    582673 ns/op	   18169 B/op	     415 allocs/op
BenchmarkSelectSlice/mysql    	    1981	    557829 ns/op	   21640 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    2077	    607452 ns/op	   21624 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1998	    552103 ns/op	   21632 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    2056	    552281 ns/op	   21624 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1993	    599284 ns/op	   21640 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1876	    604752 ns/op	   21624 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1828	    563945 ns/op	   21608 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    1909	    630789 ns/op	   21624 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    2077	    551325 ns/op	   21632 B/op	     523 allocs/op
BenchmarkSelectSlice/mysql    	    2133	    555113 ns/op	   21616 B/op	     523 allocs/op
BenchmarkSelectSlice/sqlite   	    3327	    414615 ns/op	   28984 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    3342	    378555 ns/op	   28888 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    3355	    362995 ns/op	   28912 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    3050	    359935 ns/op	   28936 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    3409	    373471 ns/op	   28992 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    3422	    349571 ns/op	   28912 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    3199	    351271 ns/op	   28928 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    2978	    349694 ns/op	   28968 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    2716	    374434 ns/op	   28968 B/op	    1441 allocs/op
BenchmarkSelectSlice/sqlite   	    3320	    356474 ns/op	   28936 B/op	    1441 allocs/op
BenchmarkSelectError/pg       	    3254	    405351 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3288	    409266 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3208	    376221 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3295	    383749 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3283	    375420 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3180	    396973 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3256	    444588 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3151	    377419 ns/op	    2595 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3266	    380265 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/pg       	    3164	    427712 ns/op	    2596 B/op	      12 allocs/op
BenchmarkSelectError/mysql    	    2622	    488103 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2096	    528142 ns/op	    2594 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2434	    417349 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2664	    462844 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2971	    372487 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2815	    374543 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2924	    402678 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2820	    367022 ns/op	    2593 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2833	    368904 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/mysql    	    2977	    376545 ns/op	    2592 B/op	      13 allocs/op
BenchmarkSelectError/sqlite   	  110044	     10761 ns/op	    2640 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  125571	     10717 ns/op	    2639 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  122265	      9932 ns/op	    2639 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  117100	     11075 ns/op	    2640 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  116604	     10087 ns/op	    2640 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  116888	     10308 ns/op	    2640 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  142762	     10696 ns/op	    2640 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  125464	     10307 ns/op	    2640 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  115760	     10215 ns/op	    2640 B/op	      19 allocs/op
BenchmarkSelectError/sqlite   	  119396	     10085 ns/op	    2640 B/op	      19 allocs/op
PASS
ok  	github.com/uptrace/bun/internal/dbtest	535.054s
+ benchstat bench-cgo.txt bench-purego.txt
name                old time/op    new time/op    delta
SelectOne/pg           337µs ± 1%     359µs ±12%   +6.30%  (p=0.009 n=8+10)
SelectOne/mysql        381µs ± 8%     372µs ± 4%     ~     (p=0.696 n=10+8)
SelectOne/sqlite      27.9µs ± 7%    45.3µs ±14%  +62.51%  (p=0.000 n=10+8)
SelectSlice/pg         595µs ± 5%     653µs ±17%   +9.76%  (p=0.000 n=9+10)
SelectSlice/mysql      577µs ± 9%     588µs ± 3%     ~     (p=0.497 n=10+9)
SelectSlice/sqlite     362µs ± 5%     557µs ± 7%  +53.97%  (p=0.000 n=9+10)
SelectError/pg         398µs ±12%     410µs ±10%     ~     (p=0.684 n=10+10)
SelectError/mysql      416µs ±27%     410µs ± 9%     ~     (p=0.529 n=10+10)
SelectError/sqlite    10.4µs ± 6%    14.8µs ± 6%  +42.04%  (p=0.000 n=10+10)

name                old alloc/op   new alloc/op   delta
SelectOne/pg          5.22kB ± 0%    5.22kB ± 0%     ~     (all equal)
SelectOne/mysql       5.52kB ± 0%    5.52kB ± 0%   -0.06%  (p=0.046 n=10+8)
SelectOne/sqlite      5.70kB ± 0%    5.75kB ± 0%   +0.84%  (p=0.000 n=10+9)
SelectSlice/pg        18.1kB ± 0%    18.2kB ± 0%     ~     (p=0.514 n=10+10)
SelectSlice/mysql     21.6kB ± 0%    21.6kB ± 0%     ~     (p=0.322 n=9+10)
SelectSlice/sqlite    28.9kB ± 0%    57.6kB ± 0%  +98.93%  (p=0.000 n=10+10)
SelectError/pg        2.60kB ± 0%    2.60kB ± 0%     ~     (p=0.650 n=10+10)
SelectError/mysql     2.59kB ± 0%    2.59kB ± 0%     ~     (p=0.465 n=10+10)
SelectError/sqlite    2.64kB ± 0%    2.64kB ± 0%     ~     (all equal)

name                old allocs/op  new allocs/op  delta
SelectOne/pg            16.0 ± 0%      16.0 ± 0%     ~     (all equal)
SelectOne/mysql         25.0 ± 0%      25.0 ± 0%     ~     (all equal)
SelectOne/sqlite        51.0 ± 0%      38.0 ± 0%  -25.49%  (p=0.000 n=10+10)
SelectSlice/pg           415 ± 0%       415 ± 0%     ~     (all equal)
SelectSlice/mysql        523 ± 0%       523 ± 0%     ~     (all equal)
SelectSlice/sqlite     1.44k ± 0%     1.72k ± 0%  +19.64%  (p=0.000 n=10+10)
SelectError/pg          12.0 ± 0%      12.0 ± 0%     ~     (all equal)
SelectError/mysql       13.0 ± 0%      13.0 ± 0%     ~     (all equal)
SelectError/sqlite      19.0 ± 0%      18.0 ± 0%   -5.26%  (p=0.000 n=10+10)
+ docker-compose down -v
Stopping dbtest_postgres_1 ... done
…