@vmihailenco 2021-06-30
I wonder what these line achieve:

```shell
docker-compose down -v
docker-compose up -d
```

I can't figure out how Docker is involved...